requirements.txt
fastapi
uvicorn
pandas
openai
python-dotenv
scikit-learn
openpyxl